Cleaned up code. Added missing @Override annotations. Organized imports.

+++ src/common/micdoodle8/mods/galacticraft/core/GCCoreBlockFallenMeteor.java
@@ -3,7 +3,6 @@ package micdoodle8.mods.galacticraft.core;
 import net.minecraft.src.Entity;
 import net.minecraft.src.EntityLiving;
 import net.minecraft.src.Material;
-import net.minecraft.src.MathHelper;
 import net.minecraft.src.World;
 
 public class GCCoreBlockFallenMeteor extends GCCoreBlock
@@ -34,11 +33,11 @@ public class GCCoreBlockFallenMeteor extends GCCoreBlock
 		{
 			EntityLiving livingEntity = (EntityLiving) par5Entity;
 			
-	        par1World.playSoundEffect((double)((float)par2 + 0.5F), (double)((float)par3 + 0.5F), (double)((float)par4 + 0.5F), "random.fizz", 0.5F, 2.6F + (par1World.rand.nextFloat() - par1World.rand.nextFloat()) * 0.8F);
+	        par1World.playSoundEffect(par2 + 0.5F, par3 + 0.5F, par4 + 0.5F, "random.fizz", 0.5F, 2.6F + (par1World.rand.nextFloat() - par1World.rand.nextFloat()) * 0.8F);
 
 	        for (int var5 = 0; var5 < 8; ++var5)
 	        {
-	            par1World.spawnParticle("largesmoke", (double)par2 + Math.random(), (double)par3 + 0.2D + Math.random(), (double)par4 + Math.random(), 0.0D, 0.0D, 0.0D);
+	            par1World.spawnParticle("largesmoke", par2 + Math.random(), par3 + 0.2D + Math.random(), par4 + Math.random(), 0.0D, 0.0D, 0.0D);
 	        }
 	        
 	        if (!livingEntity.isBurning())
