+++ src/common/micdoodle8/mods/galacticraft/core/GCCoreUtil.java
@@ -420,7 +420,7 @@ public class GCCoreUtil
 	public static float calculateMarsAngleFromOtherPlanet(long par1, float par3)
 	{
         int var4 = (int)(par1 % 48000L);
-        float var5 = ((float)var4 + par3) / 48000.0F - 0.25F;
+        float var5 = (var4 + par3) / 48000.0F - 0.25F;
 
         if (var5 < 0.0F)
         {
@@ -433,7 +433,7 @@ public class GCCoreUtil
         }
 
         float var6 = var5;
-        var5 = 1.0F - (float)((Math.cos((double)var5 * Math.PI) + 1.0D) / 2.0D);
+        var5 = 1.0F - (float)((Math.cos(var5 * Math.PI) + 1.0D) / 2.0D);
         var5 = var6 + (var5 - var6) / 3.0F;
         return var5;
 	}
@@ -441,7 +441,7 @@ public class GCCoreUtil
 	public static float calculateEarthAngleFromOtherPlanet(long par1, float par3)
 	{
         int var4 = (int)(par1 % 48000L);
-        float var5 = ((float)var4 + par3) / 48000.0F - 0.25F;
+        float var5 = (var4 + par3) / 48000.0F - 0.25F;
 
         if (var5 < 0.0F)
         {
@@ -454,7 +454,7 @@ public class GCCoreUtil
         }
 
         float var6 = var5;
-        var5 = 1.0F - (float)((Math.cos((double)var5 * Math.PI) + 1.0D) / 2.0D);
+        var5 = 1.0F - (float)((Math.cos(var5 * Math.PI) + 1.0D) / 2.0D);
         var5 = var6 + (var5 - var6) / 3.0F;
         return var5;
 	}
