+++ src/common/micdoodle8/mods/galacticraft/mars/GCMarsCaveGen2.java
@@ -4,7 +4,6 @@ import java.util.Random;
 
 import net.minecraft.src.Block;
 import net.minecraft.src.IChunkProvider;
-import net.minecraft.src.MapGenBase;
 import net.minecraft.src.MathHelper;
 import net.minecraft.src.World;
 
@@ -37,8 +36,8 @@ public class GCMarsCaveGen2
         {
             for (int var12 = par4 - var6; var12 <= par4 + var6; ++var12)
             {
-                long var13 = (long)var11 * var7;
-                long var15 = (long)var12 * var9;
+                long var13 = var11 * var7;
+                long var15 = var12 * var9;
                 this.rand.setSeed(var13 ^ var15 ^ par2World.getSeed());
                 this.recursiveGenerate(par2World, var11, var12, par3, par4, par5ArrayOfByte);
             }
@@ -55,8 +54,8 @@ public class GCMarsCaveGen2
      */
     protected void generateCaveNode(long par1, int par3, int par4, int[] par5ArrayOfByte, double par6, double par8, double par10, float par12, float par13, float par14, int par15, int par16, double par17)
     {
-        double var19 = (double)(par3 * 16 + 8);
-        double var21 = (double)(par4 * 16 + 8);
+        double var19 = par3 * 16 + 8;
+        double var21 = par4 * 16 + 8;
         float var23 = 0.0F;
         float var24 = 0.0F;
         Random var25 = new Random(par1);
@@ -79,13 +78,13 @@ public class GCMarsCaveGen2
 
         for (boolean var28 = var25.nextInt(6) == 0; par15 < par16; ++par15)
         {
-            double var29 = 1.5D + (double)(MathHelper.sin((float)par15 * (float)Math.PI / (float)par16) * par12 * 1.0F);
+            double var29 = 1.5D + MathHelper.sin(par15 * (float)Math.PI / par16) * par12 * 1.0F;
             double var31 = var29 * par17;
             float var33 = MathHelper.cos(par14);
             float var34 = MathHelper.sin(par14);
-            par6 += (double)(MathHelper.cos(par13) * var33);
-            par8 += (double)var34;
-            par10 += (double)(MathHelper.sin(par13) * var33);
+            par6 += MathHelper.cos(par13) * var33;
+            par8 += var34;
+            par10 += MathHelper.sin(par13) * var33;
 
             if (var28)
             {
@@ -114,8 +113,8 @@ public class GCMarsCaveGen2
             {
                 double var35 = par6 - var19;
                 double var37 = par10 - var21;
-                double var39 = (double)(par16 - par15);
-                double var41 = (double)(par12 + 2.0F + 16.0F);
+                double var39 = par16 - par15;
+                double var41 = par12 + 2.0F + 16.0F;
 
                 if (var35 * var35 + var37 * var37 - var39 * var39 > var41 * var41)
                 {
@@ -193,11 +192,11 @@ public class GCMarsCaveGen2
                     {
                         for (var42 = var55; var42 < var36; ++var42)
                         {
-                            double var59 = ((double)(var42 + par3 * 16) + 0.5D - par6) / var29;
+                            double var59 = (var42 + par3 * 16 + 0.5D - par6) / var29;
 
                             for (var45 = var56; var45 < var40; ++var45)
                             {
-                                double var46 = ((double)(var45 + par4 * 16) + 0.5D - par10) / var29;
+                                double var46 = (var45 + par4 * 16 + 0.5D - par10) / var29;
                                 int var48 = (var42 * 16 + var45) * 128 + var38;
                                 boolean var49 = false;
 
@@ -205,7 +204,7 @@ public class GCMarsCaveGen2
                                 {
                                     for (int var50 = var38 - 1; var50 >= var57; --var50)
                                     {
-                                        double var51 = ((double)var50 + 0.5D - par8) / var31;
+                                        double var51 = (var50 + 0.5D - par8) / var31;
 
                                         if (var51 > -0.7D && var59 * var59 + var51 * var51 + var46 * var46 < 1.0D)
                                         {
@@ -264,9 +263,9 @@ public class GCMarsCaveGen2
 
         for (int var8 = 0; var8 < var7; ++var8)
         {
-            double var9 = (double)(par2 * 16 + this.rand.nextInt(16));
-            double var11 = (double)this.rand.nextInt(this.rand.nextInt(120) + 8);
-            double var13 = (double)(par3 * 16 + this.rand.nextInt(16));
+            double var9 = par2 * 16 + this.rand.nextInt(16);
+            double var11 = this.rand.nextInt(this.rand.nextInt(120) + 8);
+            double var13 = par3 * 16 + this.rand.nextInt(16);
             int var15 = 1;
 
             if (this.rand.nextInt(4) == 0)
