+++ src/common/micdoodle8/mods/galacticraft/mars/GCMarsChunkProvider.java
@@ -22,10 +22,7 @@ import net.minecraft.src.MapGenVillage;
 import net.minecraft.src.MathHelper;
 import net.minecraft.src.NoiseGeneratorOctaves;
 import net.minecraft.src.SpawnListEntry;
-import net.minecraft.src.StructureBoundingBox;
 import net.minecraft.src.World;
-import net.minecraft.src.WorldGenLakes;
-import cpw.mods.fml.common.FMLLog;
 
 /**
  * Copyright 2012, micdoodle8
@@ -141,7 +138,7 @@ public class GCMarsChunkProvider extends ChunkProviderGenerate
 							{
 								if ((var47 += var49) > 0.0D) 
 								{
-									par3ArrayOfint[var43 += var44] = (int) GCMarsConfigManager.idBlockMarsStone;
+									par3ArrayOfint[var43 += var44] = GCMarsConfigManager.idBlockMarsStone;
 								}
 								else if (var12 * 8 + var31 < var6) 
 								{
@@ -181,8 +178,8 @@ public class GCMarsChunkProvider extends ChunkProviderGenerate
 				float var11 = var10.getFloatTemperature();
 				int var12 = (int) (this.stoneNoise[var8 + var9 * 16] / 3.0D + 3.0D + this.rand.nextDouble() * 0.25D);
 				int var13 = -1;
-				int var14 = (int) GCMarsConfigManager.idBlockMarsGrass;
-				int var15 = (int) GCMarsConfigManager.idBlockMarsDirt;
+				int var14 = GCMarsConfigManager.idBlockMarsGrass;
+				int var15 = GCMarsConfigManager.idBlockMarsDirt;
 
 				for (int var16 = 127; var16 >= 0; --var16) 
 				{
@@ -190,7 +187,7 @@ public class GCMarsChunkProvider extends ChunkProviderGenerate
 
 					if (var16 <= 0 + this.rand.nextInt(5)) 
 					{
-						par3ArrayOfint[var17] = (int) Block.bedrock.blockID;
+						par3ArrayOfint[var17] = Block.bedrock.blockID;
 					} else 
 					{
 						int var18 = par3ArrayOfint[var17];
@@ -199,14 +196,14 @@ public class GCMarsChunkProvider extends ChunkProviderGenerate
 						{
 							var13 = -1;
 						} 
-						else if (var18 == (int) GCMarsConfigManager.idBlockMarsStone)
+						else if (var18 == GCMarsConfigManager.idBlockMarsStone)
 						{
 							if (var13 == -1)
 							{
 								if (var12 <= 0)
 								{
 									var14 = 0;
-									var15 = (int) GCMarsConfigManager.idBlockMarsStone;
+									var15 = GCMarsConfigManager.idBlockMarsStone;
 								} 
 								else if (var16 >= var5 - -16 && var16 <= var5 + 1)
 								{
@@ -218,11 +215,11 @@ public class GCMarsChunkProvider extends ChunkProviderGenerate
 								{
 									if (var11 < 0.15F)
 									{
-										var14 = (int) Block.ice.blockID;
+										var14 = Block.ice.blockID;
 									}
 									else 
 									{
-										var14 = (int) 0;
+										var14 = 0;
 									}
 								}
 
@@ -244,7 +241,7 @@ public class GCMarsChunkProvider extends ChunkProviderGenerate
 								if (var13 == 0 && var15 == Block.sand.blockID)
 								{
 									var13 = this.rand.nextInt(4);
-									var15 = (int) Block.sandStone.blockID;
+									var15 = Block.sandStone.blockID;
 								}
 							}
 						}
