+++ src/common/micdoodle8/mods/galacticraft/moon/GCMoonWorldChunkManager.java
@@ -93,7 +93,7 @@ public class GCMoonWorldChunkManager extends WorldChunkManager
 
         for (int var7 = 0; var7 < par4 * par5; ++var7)
         {
-            float var8 = (float)BiomeGenBase.biomeList[var6[var7]].getIntTemperature() / 65536.0F;
+            float var8 = BiomeGenBase.biomeList[var6[var7]].getIntTemperature() / 65536.0F;
 
             if (var8 > 1.0F)
             {
